restrucutred code
