+++ CAMPAIGNS/olympex/nav_to_czml.py
@@ -3,187 +3,30 @@ nav_to_czml takes in nav data from various aricrafts (er-2 and dc-8) of olympex 
 The generated czml can be used to plot the flight track in the CESIUM.
 """
 
-import numpy as np
-from copy import deepcopy
-import json
 import boto3
 import os
 
-model = {
-    "id": "Flight Track",
-    "name": "ER2",
-    "availability": "{}/{}",
-    "model": {
-        "gltf": "https://s3.amazonaws.com/visage-czml/iphex_HIWRAP/img/er2.gltf",
-        "scale": 100.0,
-        "minimumPixelSize": 32,
-        "maximumScale": 150.0
-    },
-    "position": {
-        "cartographicDegrees": []
-    },
-    "path": {
-        "material": {
-            "solidColor": {
-                "color": {
-                    "rgba": [0, 255, 128, 255]
-                }
-            }
-        },
-        "width": 1,
-        "resolution": 5
-    },
-    "properties": {
-        "roll": {},
-        "pitch": {},
-        "heading": {}
-    }
-}
+from nav_reader_writer import FlightTrackCzmlWriter 
+from nav_reader_writer import FlightTrackReader
 
-czml_head = {
-    "id": "document",
-    "name": "wall czml",
-    "version": "1.0"
-}
+def data_pre_process(bucket_name="ghrc-fcx-field-campaigns-szg", field_campaign = "Olympex", input_data_dir = "instrument-raw-data", output_data_dir = "instrument-processed-data", instrument_name = "nav_er2"):
+    """
+    gets raw file path to s3 defined path.
+    converts it to czml.
+    puts converted file to s3 defined path.
 
-
-class FlightTrackCzmlWriter:
-
-    def __init__(self, length):
-        self.model = deepcopy(model)
-        self.length = length
-        self.model['position']['cartographicDegrees'] = [0] * 4 * length
-        self.model['properties']['roll']['number'] = [0] * 2 * length
-        self.model['properties']['pitch']['number'] = [0] * 2 * length
-        self.model['properties']['heading']['number'] = [0] * 2 * length
-
-    def set_time(self, time_window, time_steps):
-        epoch = time_window[0]
-        end = time_window[1]
-        self.model['availability'] = "{}/{}".format(epoch, end)
-        self.model['position']['epoch'] = epoch
-        self.model['position']['cartographicDegrees'][0::4] = time_steps
-        self.model['properties']['roll']['epoch'] = epoch
-        self.model['properties']['pitch']['epoch'] = epoch
-        self.model['properties']['heading']['epoch'] = epoch
-        self.model['properties']['roll']['number'][0::2] = time_steps
-        self.model['properties']['pitch']['number'][0::2] = time_steps
-        self.model['properties']['heading']['number'][0::2] = time_steps
-
-    def set_position(self, longitude, latitude, altitude):
-        self.model['position']['cartographicDegrees'][1::4] = longitude
-        self.model['position']['cartographicDegrees'][2::4] = latitude
-        self.model['position']['cartographicDegrees'][3::4] = altitude
-
-    def set_orientation(self, roll, pitch, heading):
-        self.model['properties']['roll']['number'][1::2] = roll
-        self.model['properties']['pitch']['number'][1::2] = pitch
-        self.model['properties']['heading']['number'][1::2] = heading
-
-    def set_with_df(self, df):
-        self.set_time(*self.get_time_info(df['timestamp']))
-        self.set_position(df['lon'], df['lat'], df['height_msl'])
-        self.set_orientation(df['roll'], df['pitch'], df['track'])
-
-    def get_time_info(self, time):
-        time_window = time[[0, -1]].astype(np.string_)
-        time_window = np.core.defchararray.add(time_window, np.string_('Z'))
-        time_window = np.core.defchararray.decode(time_window, 'UTF-8')
-        time_steps = (time - time[0]).astype(int)
-        return time_window, time_steps
-
-    def get_string(self):
-        return json.dumps([czml_head, self.model])
-
-
-# 1:  time
-# 2:  latitude
-# 3:  longitude
-# 4:  altitude
-# 13: heading
-# 16: pitch
-# 17: roll
-# with open('olympex_naver2_IWG1_20151109-2159.txt', newline='') as csvfile:
-class FlightTrackReader:
-
-    def __init__(self):
-        self.converters = {}
-        for i in range(33):
-            self.converters[i] = self.ignore
-        self.converters[1] = self.string_to_date
-        self.converters[2] = self.string_to_float
-        self.converters[3] = self.string_to_float
-        self.converters[4] = self.string_to_float
-        self.converters[14] = self.string_to_float
-        self.converters[16] = self.string_to_float
-        self.converters[17] = self.string_to_float
-
-    def read_csv(self, infile):
-        data = np.loadtxt(infile, delimiter=',', converters=self.converters)
-        time = data[:, 1]
-        latitude = data[:, 2]
-        longitude = data[:, 3]
-        altitude = data[:, 4]
-        heading = data[:, 14] * np.pi / 180. - np.pi / 2.
-        pitch = data[:, 16] * np.pi / 180.
-        roll = data[:, 17] * np.pi / 180.
-
-        mask = np.logical_not(np.isnan(latitude))
-        mask = np.logical_and(mask, np.logical_not(np.isnan(longitude)))
-        mask = np.logical_and(mask, np.logical_not(np.isnan(altitude)))
-        mask = np.logical_and(mask, np.logical_not(np.isnan(heading)))
-        mask = np.logical_and(mask, np.logical_not(np.isnan(pitch)))
-        mask = np.logical_and(mask, np.logical_not(np.isnan(roll)))
-
-        _, unique_idx = np.unique(time, return_index=True)
-        unique = np.copy(mask)
-        unique[:] = False
-        unique[unique_idx] = True
-
-        mask = np.logical_and(mask, unique)
-
-        time = time[mask].astype('datetime64[s]')
-        time_window = time[[0, -1]].astype(np.string_)
-        time_window = np.core.defchararray.add(time_window, np.string_('Z'))
-        self.time_window = np.core.defchararray.decode(time_window, 'UTF-8')
-        self.time_steps = (time - time[0]).astype(int).tolist()[::5]
-        self.latitude = latitude[mask][::5]
-        self.longitude = longitude[mask][::5]
-        self.altitude = altitude[mask][::5]
-        self.heading = heading[mask][::5]
-        self.pitch = pitch[mask][::5]
-        self.roll = roll[mask][::5]
-        self.length = mask[mask][::5].size
-
-    def string_to_float(self, str):
-        value = np.nan
-        try:
-            value = float(str)
-        except:
-            pass
-        return value
-
-    def string_to_date(self, str):
-        time = np.datetime64(str, 's')
-        return time.astype(np.int64)
-
-    def ignore(self, value):
-        return np.nan
-
-
-def process_tracks():
+    Args:
+        bucket_name (str, optional): source bucket. Defaults to "ghrc-fcx-field-campaigns-szg".
+        field_campaign (str, optional): name of field campaign. Case sensitive. Defaults to "Olympex".
+        input_data_dir (str, optional): folder name where raw data sits. Case sensitive. Defaults to "instrument-raw-data".
+        output_data_dir (str, optional): folder name where converted data will be stored. Case sensitive. Defaults to "instrument-processed-data".
+        instrument_name (str, optional): instrument from which data is collected. Defaults to "nav_er2".
+    """
     s3_resource = boto3.resource('s3')
-    # s3://ghrc-fcx-field-campaigns-szg/Olympex/instrument-raw-data/nav_er2/
-    bucket_name = "ghrc-fcx-field-campaigns-szg"
-    # bucket_name = os.getenv('RAW_DATA_BUCKET')
-    s3bucket = s3_resource.Bucket(bucket_name)
-    field_campaign = "Olympex"
-    input_data_dir = "instrument-raw-data"
-    output_data_dir = "instrument-processed-data"
-    instrument_name = "nav_er2"
+    s3bucket = s3_resource.Bucket(bucket_name)    
     keys = []
     for obj in s3bucket.objects.filter(
-            Prefix=f"{field_campaign}/{input_data_dir}/{instrument_name}/olympex_naver2"):
+            Prefix=f"{field_campaign}/{input_data_dir}/{instrument_name}/olympex"):
         keys.append(obj.key)
 
     result = keys
@@ -206,5 +49,24 @@ def process_tracks():
         outfile = f"{field_campaign}/{output_data_dir}/{instrument_name}/{output_name}.czml"
         s3_client.put_object(Body=output_czml, Bucket=bucket_name, Key=outfile)
 
+def er2():
+    # bucket_name = os.getenv('RAW_DATA_BUCKET')
+    bucket_name="ghrc-fcx-field-campaigns-szg"
+    field_campaign = "Olympex"
+    input_data_dir = "instrument-raw-data"
+    output_data_dir = "instrument-processed-data"
+    instrument_name = "nav_er2"
+    data_pre_process(bucket_name, field_campaign, input_data_dir, output_data_dir, instrument_name)
 
-process_tracks()
+def dc8():
+    # bucket_name = os.getenv('RAW_DATA_BUCKET')
+    bucket_name="ghrc-fcx-field-campaigns-szg"
+    field_campaign = "Olympex"
+    input_data_dir = "instrument-raw-data"
+    output_data_dir = "instrument-processed-data"
+    instrument_name = "nav_dc8"
+    data_pre_process(bucket_name, field_campaign, input_data_dir, output_data_dir, instrument_name)
+    
+er2()
+# dc8()
+
